Fix compiler warnings and add a .gitignore (#3)

+++ Sources/SwiftMockGenCore/Utils/ScannerUtils.swift
@@ -112,7 +112,7 @@ extension Structure {
         }
     }
     
-    func extract(offset: Int64, startOffset: Int64 = 0, length: Int64, content: String) -> String {
+    func extract(offset: Int, startOffset: Int = 0, length: Int, content: String) -> String {
         let end = offset + length
         let start = offset + startOffset
         
@@ -132,8 +132,8 @@ extension Structure {
     
     // This extracts the body of this structure, i.e. it doens't include the decl or signature
     func extractBody(_ file: String) -> String {
-        let start = dictionary["key.bodyoffset"] as? Int64 ?? -1
-        let len = dictionary["key.bodylength"] as? Int64 ?? 0
+        let start = dictionary["key.bodyoffset"] as? Int ?? -1
+        let len = dictionary["key.bodylength"] as? Int ?? 0
         return extract(offset: start, length: len, content: file)
     }
 }
